Update auto-generator.ts for extract result
When I use this great tool, I meet an issue.

in use of postgresql and sequelize v6, for the type of timestamp, I got Sequelize.Sequelize.literal('CURRENT_TIMESTAMP') in the auto-generated source. It's worked but I got Date only result in the postgresql.

after a little try, I notice that the generated source should tweak like:

Sequelize.Sequelize.literal('CURRENT_TIMESTAMP')
to
sequelizze.literal('CURRENT_TIMESTAMP')

use this, I got the truly timestamp with time zone result for me.

so I change the source a little, please make more comprehensive test for this.

and thanks for the great tool again!